fixed edit email bug and get/put endpoints
