+++ src/components/ProfileModal/ProfileModal.tsx
@@ -32,19 +32,17 @@ export function ProfileModal() {
 
   useEffect(() => {
     if (currentUser) {
-      getUserData(String(currentUser.email));
+      getUserData();
     }
   }, [currentUser]);
 
-  async function getUserData(email: string) {
+  async function getUserData() {
     try {
-      const params = new URLSearchParams({ email: email });
-      const response = await fetch(`/api/users?${params.toString()}`);
+      const response = await fetch(`/api/users/${currentUser?._id}`);
       const data = await response.json();
       setUserData(data.data);
     } catch (error) {
       console.error("Error getting user:", error);
-      throw error;
     }
   }
 
@@ -55,7 +53,7 @@ export function ProfileModal() {
     try {
       const params = new URLSearchParams({ type: type });
 
-      const response = await fetch(`/api/users?${params.toString()}`, {
+      const response = await fetch(`/api/users/${currentUser?._id}/?${params.toString()}`, {
         method: "PUT",
         headers: {
           "Content-Type": "application/json",
