+++ src/pages/api/users/index.ts
@@ -1,9 +1,6 @@
 import { z } from "zod";
 import {
-  createUser,
-  editUser,
-  getUser,
-  editPassword,
+  createUser
 } from "../../../server/db/actions/UserAction";
 import { userSchema } from "../../../utils/types";
 import { NextApiRequest, NextApiResponse } from "next";
@@ -12,14 +9,10 @@ import {
   HTTP_CREATED,
   HTTP_INTERNAL_SERVER_ERROR,
   HTTP_METHOD_NOT_ALLOWED,
-  HTTP_NOT_FOUND,
-  HTTP_UNAUTHORIZED,
+
 } from "@/utils/consts";
 import {
-  GenericUserErrorException,
-  UserAlreadyExistsException,
-  UserCredentialsIncorrectException,
-  UserDoesNotExistException,
+  UserAlreadyExistsException
 } from "@/utils/exceptions";
 
 export const createUserSchema = userSchema
@@ -35,31 +28,6 @@ export default async function handler(
   switch (req.method) {
     case "POST":
       await createUserHandler(req, res);
-      break;
-
-    case "GET":
-      const email = String(req.query.email);
-      try {
-        const user = await getUser(email);
-        res.status(200).send({
-          data: user,
-        });
-        return;
-      } catch (e: unknown) {
-        if (e instanceof UserDoesNotExistException) {
-          res.status(HTTP_NOT_FOUND).json({
-            error: (e as Error).message,
-          });
-          return;
-        }
-        res.status(HTTP_INTERNAL_SERVER_ERROR).json({
-          error: (e as Error).message,
-        });
-        return;
-      }
-
-    case "PUT":
-      await editUserHandler(req, res);
       break;
 
     default:
@@ -94,62 +62,5 @@ async function createUserHandler(req: NextApiRequest, res: NextApiResponse) {
       error: (e as Error).message,
     });
     return;
-  }
-}
-
-async function editUserHandler(req: NextApiRequest, res: NextApiResponse) {
-  const { type } = req.query;
-  if (type === "info") {
-    // Editing user profile
-    try {
-      const result = await editUser(req.body);
-      res.status(200).send({
-        data: result,
-      });
-      return;
-    } catch (e: unknown) {
-      if (e instanceof GenericUserErrorException) {
-        res.status(HTTP_BAD_REQUEST).json({
-          error: "User with email already exists",
-        });
-        return;
-      }
-      if (e instanceof UserDoesNotExistException) {
-        res.status(HTTP_NOT_FOUND).json({
-          error: (e as Error).message,
-        });
-        return;
-      }
-      res.status(HTTP_INTERNAL_SERVER_ERROR).json({
-        error: (e as Error).message,
-      });
-      return;
-    }
-  } else if (type === "password") {
-    // Editing user password
-    try {
-      const result = await editPassword(req.body);
-      res.status(200).send({
-        data: result,
-      });
-      return;
-    } catch (e: unknown) {
-      if (e instanceof UserDoesNotExistException) {
-        res.status(HTTP_NOT_FOUND).json({
-          error: (e as Error).message,
-        });
-        return;
-      }
-      if (e instanceof UserCredentialsIncorrectException) {
-        res.status(HTTP_UNAUTHORIZED).json({
-          error: "Old password is incorrect",
-        });
-        return;
-      }
-      res.status(HTTP_INTERNAL_SERVER_ERROR).json({
-        error: (e as Error).message,
-      });
-      return;
-    }
   }
 }
